Add default read and write capacity for all tables
If a table (or its indices) has autoscaling capacity, DynamoDB's
describe-tables operation will report zero read and write capacity.
This works, until you have to restore the table, at which point it
will complain that you can't specify zero as a read or write capacity
for a table or index. This commit sets such capacities to a default
of 25 units.